+++ app/lib/modules/transactions_module/transactions_routes.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../core/routing/routing.dart';
+import 'domain/entities/transaction_direction.dart';
 import 'presentation/new_transaction/new_transaction_page.dart';
 import 'presentation/transactions_list/transactions_list_page.dart';
 
@@ -10,6 +11,7 @@ abstract final class TransactionsRoutes {
   static const path = '/transacoes';
   static const newTransactionName = 'transactions-new';
   static const newTransactionPath = 'nova';
+  static const newTransactionDirectionQueryParameter = 'direcao';
 
   static GoRoute get route => GoRoute(
     path: path,
@@ -31,4 +33,14 @@ abstract final class TransactionsRoutes {
 
   static Future<bool?> pushNewTransactionNamed(BuildContext context) =>
       context.pushNamed<bool>(newTransactionName);
+
+  static Future<bool?> pushNewTransactionWithDirectionNamed(
+    BuildContext context,
+    TransactionDirection direction,
+  ) => context.pushNamed<bool>(
+    newTransactionName,
+    queryParameters: {
+      newTransactionDirectionQueryParameter: direction.wireValue,
+    },
+  );
 }